Show grid visualization when child is selected
